Crate onig_sys

Structs

OnigCaptureTreeNode

OnigCaseFoldCodeItem

OnigCompileInfo

OnigEncodingType

OnigErrorInfo

OnigMetaCharTableType

OnigRegexType

OnigRegion

OnigRepeatRange

OnigSyntaxType

Constants

INTERNAL_ONIGENC_CASE_FOLD_MULTI_CHAR

ONIGENC_CASE_FOLD_MIN

ONIGENC_CASE_FOLD_TURKISH_AZERI

ONIGENC_CODE_TO_MBC_MAXLEN

ONIGENC_CTYPE_ALNUM

ONIGENC_CTYPE_ALPHA

ONIGENC_CTYPE_ASCII

ONIGENC_CTYPE_BLANK

ONIGENC_CTYPE_CNTRL

ONIGENC_CTYPE_DIGIT

ONIGENC_CTYPE_GRAPH

ONIGENC_CTYPE_LOWER

ONIGENC_CTYPE_NEWLINE

ONIGENC_CTYPE_PRINT

ONIGENC_CTYPE_PUNCT

ONIGENC_CTYPE_SPACE

ONIGENC_CTYPE_UPPER

ONIGENC_CTYPE_WORD

ONIGENC_CTYPE_XDIGIT

ONIGENC_GET_CASE_FOLD_CODES_MAX_NUM

ONIGENC_MAX_COMP_CASE_FOLD_CODE_LEN

ONIGENC_MAX_STD_CTYPE

ONIGENC_MBC_CASE_FOLD_MAXLEN

ONIGERR_CHAR_CLASS_VALUE_AT_END_OF_RANGE

ONIGERR_CHAR_CLASS_VALUE_AT_START_OF_RANGE

ONIGERR_CONTROL_CODE_SYNTAX

ONIGERR_DEFAULT_ENCODING_IS_NOT_SETTED

ONIGERR_EMPTY_CHAR_CLASS

ONIGERR_EMPTY_GROUP_NAME

ONIGERR_EMPTY_RANGE_IN_CHAR_CLASS

ONIGERR_END_PATTERN_AT_CONTROL

ONIGERR_END_PATTERN_AT_ESCAPE

ONIGERR_END_PATTERN_AT_LEFT_BRACE

ONIGERR_END_PATTERN_AT_LEFT_BRACKET

ONIGERR_END_PATTERN_AT_META

ONIGERR_END_PATTERN_IN_GROUP

ONIGERR_END_PATTERN_WITH_UNMATCHED_PARENTHESIS

ONIGERR_GROUP_NUMBER_OVER_FOR_CAPTURE_HISTORY

ONIGERR_INVALID_ARGUMENT

ONIGERR_INVALID_BACKREF

ONIGERR_INVALID_CHAR_IN_GROUP_NAME

ONIGERR_INVALID_CHAR_PROPERTY_NAME

ONIGERR_INVALID_CODE_POINT_VALUE

ONIGERR_INVALID_COMBINATION_OF_OPTIONS

ONIGERR_INVALID_GROUP_NAME

ONIGERR_INVALID_LOOK_BEHIND_PATTERN

ONIGERR_INVALID_POSIX_BRACKET_TYPE

ONIGERR_INVALID_REPEAT_RANGE_PATTERN

ONIGERR_INVALID_WIDE_CHAR_VALUE

ONIGERR_MATCH_STACK_LIMIT_OVER

ONIGERR_MEMORY

ONIGERR_META_CODE_SYNTAX

ONIGERR_MISMATCH_CODE_LENGTH_IN_CLASS_RANGE

ONIGERR_MULTIPLEX_DEFINED_NAME

ONIGERR_MULTIPLEX_DEFINITION_NAME_CALL

ONIGERR_NESTED_REPEAT_OPERATOR

ONIGERR_NEVER_ENDING_RECURSION

ONIGERR_NOT_SUPPORTED_ENCODING_COMBINATION

ONIGERR_NUMBERED_BACKREF_OR_CALL_NOT_ALLOWED

ONIGERR_OVER_THREAD_PASS_LIMIT_COUNT

ONIGERR_PARSER_BUG

ONIGERR_PREMATURE_END_OF_CHAR_CLASS

ONIGERR_SPECIFIED_ENCODING_CANT_CONVERT_TO_WIDE_CHAR

ONIGERR_STACK_BUG

ONIGERR_TARGET_OF_REPEAT_OPERATOR_INVALID

ONIGERR_TARGET_OF_REPEAT_OPERATOR_NOT_SPECIFIED

ONIGERR_TOO_BIG_BACKREF_NUMBER

ONIGERR_TOO_BIG_NUMBER

ONIGERR_TOO_BIG_NUMBER_FOR_REPEAT_RANGE

ONIGERR_TOO_BIG_WIDE_CHAR_VALUE

ONIGERR_TOO_LONG_WIDE_CHAR_VALUE

ONIGERR_TOO_MANY_MULTI_BYTE_RANGES

ONIGERR_TOO_SHORT_MULTI_BYTE_STRING

ONIGERR_TYPE_BUG

ONIGERR_UNDEFINED_BYTECODE

ONIGERR_UNDEFINED_GROUP_OPTION

ONIGERR_UNDEFINED_GROUP_REFERENCE

ONIGERR_UNDEFINED_NAME_REFERENCE

ONIGERR_UNEXPECTED_BYTECODE

ONIGERR_UNMATCHED_CLOSE_PARENTHESIS

ONIGERR_UNMATCHED_RANGE_SPECIFIER_IN_CHAR_CLASS

ONIGERR_UPPER_SMALLER_THAN_LOWER_IN_REPEAT_RANGE

ONIG_CHAR_TABLE_SIZE

ONIG_INEFFECTIVE_META_CHAR

ONIG_INFINITE_DISTANCE

ONIG_MAX_BACKREF_NUM

ONIG_MAX_CAPTURE_HISTORY_GROUP

ONIG_MAX_ERROR_MESSAGE_LEN

ONIG_MAX_MULTI_BYTE_RANGES_NUM

ONIG_MAX_REPEAT_NUM

ONIG_META_CHAR_ANYCHAR

ONIG_META_CHAR_ANYCHAR_ANYTIME

ONIG_META_CHAR_ANYTIME

ONIG_META_CHAR_ESCAPE

ONIG_META_CHAR_ONE_OR_MORE_TIME

ONIG_META_CHAR_ZERO_OR_ONE_TIME

ONIG_MISMATCH

ONIG_NORMAL

ONIG_NO_SUPPORT_CONFIG

ONIG_NREGION

ONIG_OPTION_CAPTURE_GROUP

ONIG_OPTION_DEFAULT

ONIG_OPTION_DONT_CAPTURE_GROUP

ONIG_OPTION_EXTEND

ONIG_OPTION_FIND_LONGEST

ONIG_OPTION_FIND_NOT_EMPTY

ONIG_OPTION_IGNORECASE

ONIG_OPTION_MAXBIT

ONIG_OPTION_MULTILINE

ONIG_OPTION_NEGATE_SINGLELINE

ONIG_OPTION_NONE

ONIG_OPTION_NOTBOL

ONIG_OPTION_NOTEOL

ONIG_OPTION_POSIX_REGION

ONIG_OPTION_SINGLELINE

ONIG_REGION_NOTPOS

ONIG_STATE_COMPILING

ONIG_STATE_MODIFY

ONIG_STATE_NORMAL

ONIG_STATE_SEARCHING

ONIG_SYN_ALLOW_DOUBLE_RANGE_OP_IN_CC

ONIG_SYN_ALLOW_EMPTY_RANGE_IN_CC

ONIG_SYN_ALLOW_INTERVAL_LOW_ABBREV

ONIG_SYN_ALLOW_INVALID_INTERVAL

ONIG_SYN_ALLOW_MULTIPLEX_DEFINITION_NAME

ONIG_SYN_ALLOW_UNMATCHED_CLOSE_SUBEXP

ONIG_SYN_BACKSLASH_ESCAPE_IN_CC

ONIG_SYN_CAPTURE_ONLY_NAMED_GROUP

ONIG_SYN_CONTEXT_INDEP_ANCHORS

ONIG_SYN_CONTEXT_INDEP_REPEAT_OPS

ONIG_SYN_CONTEXT_INVALID_REPEAT_OPS

ONIG_SYN_DIFFERENT_LEN_ALT_LOOK_BEHIND

ONIG_SYN_FIXED_INTERVAL_IS_GREEDY_ONLY

ONIG_SYN_NOT_NEWLINE_IN_NEGATIVE_CC

ONIG_SYN_OP2_ATMARK_CAPTURE_HISTORY

ONIG_SYN_OP2_CCLASS_SET_OP

ONIG_SYN_OP2_ESC_CAPITAL_C_BAR_CONTROL

ONIG_SYN_OP2_ESC_CAPITAL_M_BAR_META

ONIG_SYN_OP2_ESC_CAPITAL_Q_QUOTE

ONIG_SYN_OP2_ESC_GNU_BUF_ANCHOR

ONIG_SYN_OP2_ESC_G_SUBEXP_CALL

ONIG_SYN_OP2_ESC_H_XDIGIT

ONIG_SYN_OP2_ESC_K_NAMED_BACKREF

ONIG_SYN_OP2_ESC_P_BRACE_CHAR_PROPERTY

ONIG_SYN_OP2_ESC_P_BRACE_CIRCUMFLEX_NOT

ONIG_SYN_OP2_ESC_U_HEX4

ONIG_SYN_OP2_ESC_V_VTAB

ONIG_SYN_OP2_INEFFECTIVE_ESCAPE

ONIG_SYN_OP2_OPTION_PERL

ONIG_SYN_OP2_OPTION_RUBY

ONIG_SYN_OP2_PLUS_POSSESSIVE_INTERVAL

ONIG_SYN_OP2_PLUS_POSSESSIVE_REPEAT

ONIG_SYN_OP2_QMARK_GROUP_EFFECT

ONIG_SYN_OP2_QMARK_LT_NAMED_GROUP

ONIG_SYN_OP_ASTERISK_ZERO_INF

ONIG_SYN_OP_BRACE_INTERVAL

ONIG_SYN_OP_BRACKET_CC

ONIG_SYN_OP_DECIMAL_BACKREF

ONIG_SYN_OP_DOT_ANYCHAR

ONIG_SYN_OP_ESC_ASTERISK_ZERO_INF

ONIG_SYN_OP_ESC_AZ_BUF_ANCHOR

ONIG_SYN_OP_ESC_BRACE_INTERVAL

ONIG_SYN_OP_ESC_B_WORD_BOUND

ONIG_SYN_OP_ESC_CAPITAL_G_BEGIN_ANCHOR

ONIG_SYN_OP_ESC_CONTROL_CHARS

ONIG_SYN_OP_ESC_C_CONTROL

ONIG_SYN_OP_ESC_D_DIGIT

ONIG_SYN_OP_ESC_LPAREN_SUBEXP

ONIG_SYN_OP_ESC_LTGT_WORD_BEGIN_END

ONIG_SYN_OP_ESC_OCTAL3

ONIG_SYN_OP_ESC_PLUS_ONE_INF

ONIG_SYN_OP_ESC_QMARK_ZERO_ONE

ONIG_SYN_OP_ESC_S_WHITE_SPACE

ONIG_SYN_OP_ESC_VBAR_ALT

ONIG_SYN_OP_ESC_W_WORD

ONIG_SYN_OP_ESC_X_BRACE_HEX8

ONIG_SYN_OP_ESC_X_HEX2

ONIG_SYN_OP_LINE_ANCHOR

ONIG_SYN_OP_LPAREN_SUBEXP

ONIG_SYN_OP_PLUS_ONE_INF

ONIG_SYN_OP_POSIX_BRACKET

ONIG_SYN_OP_QMARK_NON_GREEDY

ONIG_SYN_OP_QMARK_ZERO_ONE

ONIG_SYN_OP_VARIABLE_META_CHARACTERS

ONIG_SYN_OP_VBAR_ALT

ONIG_SYN_STRICT_CHECK_BACKREF

ONIG_SYN_WARN_CC_OP_NOT_ESCAPED

ONIG_SYN_WARN_REDUNDANT_NESTED_REPEAT

ONIG_TRAVERSE_CALLBACK_AT_BOTH

ONIG_TRAVERSE_CALLBACK_AT_FIRST

ONIG_TRAVERSE_CALLBACK_AT_LAST

Statics

OnigDefaultSyntax

OnigEncodingASCII

OnigEncodingBIG5

OnigEncodingCP1251

OnigEncodingEUC_CN

OnigEncodingEUC_JP

OnigEncodingEUC_KR

OnigEncodingEUC_TW

OnigEncodingGB18030

OnigEncodingISO_8859_1

OnigEncodingISO_8859_10

OnigEncodingISO_8859_11

OnigEncodingISO_8859_13

OnigEncodingISO_8859_14

OnigEncodingISO_8859_15

OnigEncodingISO_8859_16

OnigEncodingISO_8859_2

OnigEncodingISO_8859_3

OnigEncodingISO_8859_4

OnigEncodingISO_8859_5

OnigEncodingISO_8859_6

OnigEncodingISO_8859_7

OnigEncodingISO_8859_8

OnigEncodingISO_8859_9

OnigEncodingKOI8

OnigEncodingKOI8_R

OnigEncodingSJIS

OnigEncodingUTF16_BE

OnigEncodingUTF16_LE

OnigEncodingUTF32_BE

OnigEncodingUTF32_LE

OnigEncodingUTF8

OnigSyntaxASIS

OnigSyntaxEmacs

OnigSyntaxGnuRegex

OnigSyntaxGrep

OnigSyntaxJava

OnigSyntaxPerl

OnigSyntaxPerl_NG

OnigSyntaxPosixBasic

OnigSyntaxPosixExtended

OnigSyntaxRuby

Functions

onig_capture_tree_traverse
Traverse and callback in capture history data tree.

onig_copy_encoding
Copy encoding.

onig_copy_syntax
Copy syntax.

onig_copyright
Return Copyright String

onig_end
The use of this library is finished.

onig_error_code_to_str
Get error message string. If this function is used for onig_new(), don't call this after the pattern argument of onig_new() is freed.

onig_foreach_name
Iterate function call for all names.

onig_free
Free memory used by regex object.

onig_free_body
Free memory used by regex object. (Except reg oneself.)

onig_get_capture_tree
Return the root node of capture history data tree.

onig_get_case_fold_flag
OnigCaseFoldType onig_get_case_fold_flag(regex_t* reg)

onig_get_default_case_fold_flag
Get default case fold flag.

onig_get_encoding
OnigEncoding onig_get_encoding(regex_t* reg)

onig_get_match_stack_limit_size
Return the maximum number of stack size. (default: 0 == unlimited)

onig_get_options
OnigOptionType onig_get_options(regex_t* reg)

onig_get_syntax
OnigSyntaxType* onig_get_syntax(regex_t* reg)

onig_get_syntax_behavior
unsigned int onig_get_syntax_behavior(OnigSyntaxType* syntax)

onig_get_syntax_op
unsigned int onig_get_syntax_op(OnigSyntaxType* syntax)

onig_get_syntax_op2
unsigned int onig_get_syntax_op2(OnigSyntaxType* syntax)

onig_get_syntax_options
OnigOptionType onig_get_syntax_options(OnigSyntaxType* syntax)

onig_init
Initialize library.

onig_initialize
Initialise Library

onig_initialize_encoding
Onig Initialise Encoding

onig_match
Match string and return result and matching region.

onig_name_to_backref_number
Return the group number corresponding to the named backref (\k). If two or more regions for the groups of the name are effective, the greatest number in it is obtained.

onig_name_to_group_numbers
Return the group number list of the name. Named subexp is defined by (?....).

onig_new
Create a regex object.

onig_new_deluxe
Create a regex object. This function is deluxe version of onig_new().

onig_new_without_alloc
Create a regex object. reg object area is not allocated in this function.

onig_noname_group_capture_is_active
Return noname group capture activity.

onig_number_of_capture_histories
Return the number of capture history defined in the pattern.

onig_number_of_captures
Return the number of capture group in the pattern.

onig_number_of_names
Return the number of names defined in the pattern. Multiple definitions of one name is counted as one.

onig_reg_init
Onig Reg Init

onig_region_clear
Clear contents of region.

onig_region_copy
Copy contents of region.

onig_region_free
Free memory used by region.

onig_region_new
Create a region.

onig_region_resize
Resize group range area of region.

onig_search
Search string and return search result and matching region.

onig_set_default_case_fold_flag
Set default case fold flag.

onig_set_default_syntax
Set default syntax.

onig_set_match_stack_limit_size
Set the maximum number of stack size.

onig_set_meta_char
Set a variable meta character to the code point value. Except for an escape character, this meta characters specification is not work, if ONIG_SYN_OP_VARIABLE_META_CHARACTERS is not effective by the syntax. (Build-in syntaxes are not effective.)

onig_set_syntax_behavior
void onig_set_syntax_behavior(OnigSyntaxType* syntax, unsigned int behavior)

onig_set_syntax_op
void onig_set_syntax_op(OnigSyntaxType* syntax, unsigned int op)

onig_set_syntax_op2
void onig_set_syntax_op2(OnigSyntaxType* syntax, unsigned int op2)

onig_set_syntax_options
void onig_set_syntax_options(OnigSyntaxType* syntax, OnigOptionType options)

onig_set_verb_warn_func
Set verbose warning function.

onig_set_warn_func
Set warning function.

onig_unicode_define_user_property
Define User Unicode Property

onig_version
Return version string. (ex. "5.0.3")

onigenc_get_default_encoding
Onigenc Get Default Encoding

onigenc_get_left_adjust_char_head
Return left-adjusted head address of a character.

onigenc_get_prev_char_head
Return previous character head address.

onigenc_get_right_adjust_char_head
Return right-adjusted head address of a character.

onigenc_get_right_adjust_char_head_with_prev
Onigenc Get Right Adjust Char Head With Prev

onigenc_init
Onigenc Init

onigenc_set_default_caseconv_table
Onigenc Set Default Case Conversion Table

onigenc_set_default_encoding
Onigenc Set Default Encoding

onigenc_step_back
Onigenc Step Back

onigenc_str_bytelen_null
Return number of bytes in the string.

onigenc_strlen
Return number of characters in the string.

onigenc_strlen_null
Return number of characters in the string.

Type Definitions

OnigApplyAllCaseFoldFunc
Apply All Case Fold Callback, see OnigEncodingType->apply_all_case_fold

OnigCaptureTreeTraverseCallback
Capture Tree Callback

OnigCaseFoldType

OnigCodePoint

OnigCtype

OnigDistance

OnigEncoding

OnigForeachNameCallback
Foreach Callback

OnigOptionType

OnigRegex

OnigRegexMut

OnigSyntaxBehavior

OnigSyntaxOp

OnigSyntaxOp2

OnigUChar

OnigWarnFunc
Warning Callback