onig_sys::onig_capture_tree_traverse

pub unsafe extern fn onig_capture_tree_traverse(region: *const OnigRegion, at: c_int, func: OnigCaptureTreeTraverseCallback, arg: c_void) -> c_int

Traverse and callback in capture history data tree.

int onig_capture_tree_traverse(OnigRegion* region, int at, int(*func)(int,int,int,int,int,void*), void* arg)

Returns

normal return: 0 error: callback func's return value.

Arguments

region: match region data.

at: callback position.

ONIG_TRAVERSE_CALLBACK_AT_FIRST: callback first, then traverse childs.

ONIG_TRAVERSE_CALLBACK_AT_LAST: traverse childs first, then callback.

ONIG_TRAVERSE_CALLBACK_AT_BOTH: callback first, then traverse childs, and at last callback again.

func: callback function. if func does not return 0, then traverse is stopped.

 int func(int group, int beg, int end, int level, int at,
          void* arg)

group: group number

beg: capture start position

end: capture end position

level: nest level (from 0)

at: callback position

ONIG_TRAVERSE_CALLBACK_AT_FIRST

ONIG_TRAVERSE_CALLBACK_AT_LAST

arg: optional callback argument

arg; optional callback argument.