onig_sys::onig_set_meta_char

pub unsafe extern fn onig_set_meta_char(syntax: *mut OnigSyntaxType, what: c_uint, code: OnigCodePoint) -> c_int

Set a variable meta character to the code point value. Except for an escape character, this meta characters specification is not work, if ONIG_SYN_OP_VARIABLE_META_CHARACTERS is not effective by the syntax. (Build-in syntaxes are not effective.)

int onig_set_meta_char(OnigSyntaxType* syntax, unsigned int what, OnigCodePoint code)

normal return: ONIG_NORMAL

arguments 1 syntax: target syntax 2 what: specifies which meta character it is.

 ONIG_META_CHAR_ESCAPE
 ONIG_META_CHAR_ANYCHAR
 ONIG_META_CHAR_ANYTIME
 ONIG_META_CHAR_ZERO_OR_ONE_TIME
 ONIG_META_CHAR_ONE_OR_MORE_TIME
 ONIG_META_CHAR_ANYCHAR_ANYTIME

3 code: meta character or ONIG_INEFFECTIVE_META_CHAR.